+++ source/EngineInterface/GenomeConstants.h
@@ -22,8 +22,6 @@ namespace Const
     auto constexpr ConstructorConstructionAngle1Pos = 3;
     auto constexpr ConstructorConstructionAngle2Pos = 4;
 
-    auto constexpr ConstructorSeparation = 2;
-
     auto constexpr CellBasicBytes = 8;
     auto constexpr NeuronBytes = 64 + 8 + 8;
     auto constexpr TransmitterBytes = 1;