+++ source/Gui/StatisticsWindow.cpp
@@ -289,7 +289,7 @@ void _StatisticsWindow::processTimelineStatistics()
 
         ImGui::TableNextRow();
         ImGui::TableSetColumnIndex(0);
-        processPlot(row++, &DataPointCollection::numDetonations, 6);
+        processPlot(row++, &DataPointCollection::numDetonations, 8);
         ImGui::TableSetColumnIndex(1);
         AlienImGui::Text("Detonations");